// src/shared/configurations/base-types.d.ts
export type BaseAppConfigurationKeys =
  | "hidden_entity_table_columns"
  | "hidden_entity_create_columns"
  | "hidden_entity_update_columns"
  | "hidden_entity_details_columns"
  | "entity_diction"
  | "entity_columns_types"
  | "theme_color"
  | "entity_fields_orders_table"
  | "entity_fields_orders_create"
  | "entity_fields_orders_update"
  | "entity_fields_orders_details"
  | "entity_crud_settings"
  | "table_views"
  | "persistent_query"
  | "users_to_database_link"
  | "menu_entities_order"
  | "disabled_menu_entities"
  | "default_date_format"
  | "entity_columns_labels"
  | "disabled_entities"
  | "entity_presentation_script"
  | "site_settings"
  | "entity_relation_template"
  | "entity_selections"
  | "entity_validations"
  | "entity_relations_labels"
  | "entity_form_extension"
  | "system_settings"
  | "hidden_entity_relations"
  | "file_upload_settings"
  | "metadata_columns"
  | "entity_relations_order";